Initial Commit
 * Add README, LICENSE, pom.xml, and .gitignore files

This commit includes the introduction of four initial files for the project. A README.md file provides a brief description of the SimpleConfigLib project. The LICENSE file includes the MIT License details for the project. An initial pom.xml is added for managing the project's build, while a .gitignore file defines the files/directories to ignore within the repository. (~AI)

Signed-off-by: Drachir000 <mra.berginski@gmail.com>